+++ curie_desktop/lib/src/models/countries.dart
@@ -1,6 +1,0 @@
-class Countries {
-  final int countryPK;
-  final String countryName;
-
-  Countries(this.countryPK, this.countryName);
-}