+++ lib/src/views/navigation_test/navigation_test_screen.dart
@@ -1,0 +1,120 @@
+import 'package:curie/src/templates/widgets/appbar_template.dart';
+import 'package:flutter/material.dart';
+
+class NavigationTestScreen extends StatefulWidget {
+  const NavigationTestScreen({Key? key}) : super(key: key);
+
+  static const routeName = '/test';
+
+  @override
+  State<NavigationTestScreen> createState() => _NavigationTestScreenState();
+}
+
+class _NavigationTestScreenState extends State<NavigationTestScreen> {
+  int _selectedIndex = 0;
+
+  @override
+  Widget build(BuildContext context) {
+    return SafeArea(
+        child: Scaffold(
+      appBar: appBarTemplate(context: context, title: 'title'),
+      body: Row(
+        children: [
+          NavigationRail(
+            elevation: 8,
+            selectedIndex: _selectedIndex,
+            onDestinationSelected: (index) {
+              setState(() {
+                _selectedIndex = index;
+              });
+            },
+            labelType: NavigationRailLabelType.all,
+            destinations: const [
+              NavigationRailDestination(
+                icon: Icon(Icons.favorite_border),
+                selectedIcon: Icon(Icons.favorite),
+                label: Text('First'),
+              ),
+              NavigationRailDestination(
+                icon: Icon(Icons.bookmark_border),
+                selectedIcon: Icon(Icons.book),
+                label: Text('Second'),
+              ),
+              NavigationRailDestination(
+                icon: Icon(Icons.star_border),
+                selectedIcon: Icon(Icons.star),
+                label: Text('Third'),
+              ),
+            ],
+          ),
+          const VerticalDivider(thickness: 1, width: 1),
+          // This is the main content.
+          Expanded(
+            child: IndexedStack(
+              index: _selectedIndex,
+              children: const [
+                FirstWidget(),
+                SecondWidget(),
+                ThirdWidget(),
+              ],
+            ),
+          ),
+        ],
+      ),
+    ));
+  }
+}
+
+class FirstWidget extends StatelessWidget {
+  const FirstWidget({Key? key}) : super(key: key);
+
+  @override
+  Widget build(BuildContext context) {
+    return Center(
+      child: Container(
+        width: 1152,
+        height: 128,
+        child: Card(
+          elevation: 2,
+          child: Padding(
+            padding: const EdgeInsets.all(16.0),
+            child: Column(
+              children: [
+                Text(
+                  'This is the first Widget',
+                  style: Theme.of(context).textTheme.headline6,
+                ),
+                Text(
+                  "Lorem Ipsum is simply dummy text of the printing and typesetting industry. Lorem Ipsum has been the industry's standard dummy text ever since the 1500s, when an unknown printer took a galley of type and scrambled it to make a type specimen book. It has survived not only five centuries, but also the leap into electronic typesetting, remaining essentially unchanged. It was popularised in the 1960s with the release of Letraset sheets containing Lorem Ipsum passages, and more recently with desktop publishing software like Aldus PageMaker including versions of Lorem Ipsum.",
+                  style: Theme.of(context).textTheme.bodyText1,
+                ),
+              ],
+            ),
+          ),
+        ),
+      ),
+    );
+  }
+}
+
+class SecondWidget extends StatelessWidget {
+  const SecondWidget({Key? key}) : super(key: key);
+
+  @override
+  Widget build(BuildContext context) {
+    return const Center(
+      child: Text('This is the second Widget'),
+    );
+  }
+}
+
+class ThirdWidget extends StatelessWidget {
+  const ThirdWidget({Key? key}) : super(key: key);
+
+  @override
+  Widget build(BuildContext context) {
+    return const Center(
+      child: Text('This is the third Widget'),
+    );
+  }
+}