+++ CHANGELOG.md
@@ -2,6 +2,12 @@
 
 All notable changes in this project will be documented in this file.
 
-## Curie Desktop v2021.1.0-Pre-Alpha (December 15, 2021)
+## Curie v2022.0.1-Pre-Alpha (August 16, 2022)
+
+- Implemented ThemeMode (dark theme and light theme).
+- Implemented Locale ('en' and 'pt' support).
+- Started development of the app.
+
+## Curie v2021.0.0-Pre-Alpha (December 15, 2021)
 
 - initial Config